Polished functions across the lka algo

+++ apps/rpi5-firmware/Lka.hpp
@@ -66,5 +66,5 @@ public:
 private:
 
 	Bev _bev;
-	int _angle;
+	float _angle;
 };
